Initial prompt commit
Initial prompt:

Can you create a nextjs website. This will be a website that let's my wedding guests rsvp to my wedding, it includes venue location, wedding time and day. An option to RSVP (attending or not) and how many guests are coming with them. Also an option to choose which meal they want (beef, or fish, or vegeterian). Each page should have a unique url based of the main guests name with a uniq identifier. You can create these individual pages based off of a json file of a list of guests (please make your own with fake data for now). Once a guests fills out their details, it will be saved to another json file with all the guests answers. Let me know if you have any questions.